+++ src/Test/ScrapeMovieTest.cs
@@ -52,5 +52,29 @@ namespace JavLuv
             Assert.NotEmpty(module.ImageSource);
         }
 
+        [Fact]
+        public void TestScrapeMovieJavLand()
+        {
+            var metadata = new MovieMetadata("IESP-711");
+            var module = new MovieJavLand(metadata, LanguageType.English);
+            module.Scrape();
+            Assert.Equal("Minami Iroha Lesbian Ban Lifted ~I Was Dragged Into A Lesbian Swamp By My Sister-In-Law ~", module.Metadata.Title);
+            Assert.Equal("2022-12-22", module.Metadata.Premiered);
+            Assert.Equal(142, module.Metadata.Runtime);
+            Assert.Equal("Chrysanthemum", module.Metadata.Director);
+            Assert.Equal("I-Energy", module.Metadata.Studio);
+            Assert.Equal("Iesp", module.Metadata.Label);
+            Assert.Equal(5, module.Metadata.Genres.Count);
+            Assert.Equal("High Vision", module.Metadata.Genres[0]);
+            Assert.Equal("Drama", module.Metadata.Genres[1]);
+            Assert.Equal("Big Tits", module.Metadata.Genres[2]);
+            Assert.Equal("Slender", module.Metadata.Genres[3]);
+            Assert.Equal("Lesbian", module.Metadata.Genres[4]);
+            Assert.Equal(2, module.Metadata.Actors.Count);
+            Assert.Equal("Iroha Minami", module.Metadata.Actors[0].Name);
+            Assert.Equal("Tadami Rumi", module.Metadata.Actors[1].Name);
+            Assert.NotEmpty(module.ImageSource);
+        }
+
     }
 }