+++ Changelog.md
@@ -2,9 +2,9 @@
 
 All significant changes are listed in this file.
 
-## [1.1.27] - 2023-4-28
+## [1.1.27] - 2023-5-6
 - Add automated tests for movie scrapers for improved future reliability
-- Fix several issues in JavDatabase actress scraper
+- Fix several issues in actress and movie scrapers
 - Add filter to correct bad data in Actress database
 
 ## [1.1.26] - 2023-4-27